Day 8: Solve day 8
Question for a moment why we are building a treehouse literally one day of travel away from the ship which took an entire week to unpack.

Tell communication device it is ugly.

Attempt to educate elves on the concept of not changing visibility rules mid-challenge.

Build anti-ladybug walls out of grass.