Day 9: Solve day 9
Question for a moment why we built a treehouse for a day but can't build decent bridges we use every year?

Tell communication device it deserved its bath

Attempt to educate elves on the concept of bridge building (but the river is very loud.)

Lament loss of now-soaked medicinal weeds.