Day 13: Solve day 13
Question for a moment what an absolute bastard would put 10 as a possible value into the input, never make any allusion to values having multiple digits, and use exclusively examples with numbers 0-9, even using a list of 1-9 but stopping at 10, necessitating implementation with known input which inevitably leads to algorithms hacked to deal with the specific gotcha traps not to do which is only the entire point of having huge randomised inputs!

Suddenly realise the elves packed no means of making fire, and the monkey death blaster is not conducive to lighting anything without annihilating it.

Attempt to educate elves telepathically on what a bastard Santa is.