Day 14: Solve day 14
Question for a moment what was the point of simulating all this sand when not getting trapped by it still involves running to the edge as fast as possible.

Carefully brush sand off of medicinal weeds stash.

Attempt to educate elves telepathically on how they can all rot and I'm going to have an adventure on my own.