Day 16: Solve day 16
Question for a moment why every challenge seems to have become "deal with this n! hard problem". But again. Seriously, we can do better than that.

Consider medicinal weeds stash, but decide against it (It would not burn properly in an erupting volcano.)

Attempt to educate elephants on how long it would take for a trillion rocks to fall at one rock a second (189k years and some.)

Pre-emptively erect shield protecting me from people saying a solution is only valid if it passes for all inputs and not coincidentally this input (as *clearly* input hacking is part of the challenge).