Day 20: Solve day 20
Question for a moment what happened to maps. Like, made of paper. (They fell in the river.)

Also question for a moment why I need to manually break the encryption when I HAVE THE DECRYPTION KEY.

Consider medicinal weeds stash, but decide against it. (I am already hungry!)

Attempt to educate elves telepathically on a little tale of Alice and Bob.