Day 21: Solve day 21.
Let's talk for a moment. Fuck the silly stories and roleplay, which have been nonsensical at best this year to string together exponential optimisation problems.

AoC is full of edge case traps. It always has been. It teaches one to be wary of them even if the specification didn't think of them. It teaches making robust software that can handle errors, or report them early when an assumption turns out to be wrong.

Nothing is wrong with THAT. This challenge, however, is one that contains a trap in a meta assumption which has been true in *every* challenge in the past four years that I have been doing these challenges: If you implement it as described, there is exactly one solution. If there are multiple, you have missed something.

Worse yet, in addition to the meta-assumption suddenly having to be ignored (AKA changing the rules), to my mind there has not been a challenge which included a trap in the specification of the behavior being incomplete.

I cannot immediately recall any challenge in which anything other than integer numbers and integer division were used. Yes, usually it is always specified what to use, and this time it wasn't. Do you know how an engineer reacts to this kind of gap in specification?

We ask. But since we cannot ask the specifier in this case, we are left to assume the default that has been the case EVERY TIME SO FAR stands, which is integer division, and that the description would note if the operation is not applicable under some conditions. Neither is the case now, and public stance I have seen is that this is supposed to be part of the challenge.

Other years I would take this as it is, but the quality of this year has been honestly disappointing in multiple challenges. What comes to mind immediately are Day 19 Part 2, whose description ends in two extremely redundant paragraphs as though one was meant to be deleted, and Day 20, which infuriatingly does not have description or an example showing what happens when a number passes over itself when finding its new spot. While the answer of removing the number first before counting steps is somewhat intuitive, you cannot know for sure until you have solved day 1, as this guess is required for the correct solution and does not come up in the example.

In short. This year's story is nonsensical. The challenges are frankly fine, but entirely too many of them boil down to searching a tree or graph, in which the algorithm itself is trivial but you need to build (real world) heuristics or exploit known input to make them tenable. But even that would be fine, if not a solid third of the Part 2s were "Now do it with way more iterations" which just isn't a challenge when most challenges are already untenable on Part 1 without heavy optimizations. And now we are left to guess at incomplete specification.

Multiple times this year have I been upset not because the problem is hard or has more edge cases than I thought, but because the challenge isn't enjoyable. Either being trivial and/or a functional repeat of another challenge from this year. This has never happened before. Why are you bringing the annoyances of professional development into our free time hacking calendar?