Day 23: Solve day 23
Question for a moment how the plants are meant to survive if they are only planted once a year and require frequent volcano eruptions in their immediate vicinity.

Consider medicinal weeds stash but decide against it. (Too much ash falling in the air.)

Attempt to educate elves on literally everything. How does the grove rely on volcanic ash if it's protected by a forcefield? Would a greenhouse have been impossible? Was no one ever surveilling this grove including the huge volcano next to it? Is their super time consuming way of finding a spot just "Be two paces away from everyone else", because that's how it ends up?