Day 25: Solve day 25
Question for a moment where I got most of the starfruit, which are hypersensitive to their environment, when I spent literally the entire time NOT IN THEIR DEDICATED ENVIRONMENT.

Also, why do the reindeer need that smoothie now and only now?

Attempt to educate elves on the whole thing with the seven fat cows and seven slim cows.

Light up a phat blunt of my gathered medicinal weeds with a little starfruit oil to forget all of this year's woes. 2022 blaze it.