Day 10: Solve day 10
Attempt to educate self on that if the metal rodent can escape to an enclosed area out of the pipes, then it can also escape to any other area because clearly they are not sealed.

Stop talking to self.

Scavenge some Definitely (TM) disconnected pipes. Copper is valuable, and if nothing else the aluminium will make paperclips.

Stash pipes in bag for later.