Day 11: Solve day 11
Attempt to educate elf on the irrelevance of this task; If the true distance between galaxies has grown million-fold between their emitting light and you observing it, the distance to those galaxies must be absurd.

Wonder how elf even observed those galaxies from so far away as space is anything but empty.

Acquire readout of elf's data. It can't be not useful eventually. Stash in bag for later.