Day 13: Solve day 13
Attempt to educate elf on that Hamming distance is a useful concept if and only if you are developing an autocorrect algorithm which seems a dying breed these days.

Remember there is no elf.

Consider simply walking with outstretched arms.

Note odd flashes of colour reflecting in field of mirrors. Disregard.