Day "14": Solve day 14
Attempt to educate empty air on the sheer impracticality of a parabolic death ray, much less whatever this is to be.

Choose not to point out that, if the rocks deform the platform, then in an analogue to gravity itself, the presence of rocks would change the path of other rocks rolling. (Still not over those damn beacons.)

Point reflector dish at self just to prove the impracticality thereof.

Notice odd pattern of light in the reflections - no, it was nothing.

Close bag.