Day "15": Solve day 15
Attempt to educate reindeer on the concept of just labelling the boxes with the focal lengths that are supposed to go into each slot.

Choose not to point out how, if such a laser focusing contraption conceptually could work, it would destroy itself as part of normal operation.

Stand in beam just to prove it. Experience a little warmth and an odd iridescent pattern.

Emerge unscathed.

Close bag.