Day "16": Solve day 16
Attempt to educate reindeer on the concept of conservation of energy; After all the effort focusing the beams, splitting them again into a glorified lighting fixture seems a bit of a waste.

Choose not to point out how this contraption is literally made to self-destruct.

Be caught in chamber by reindeer recklessly activating the machinery and submerged in rapidly melting lava.

Emerge unscathed.

Close bag.