Day "18": Solve day 18
Attempt to educate elves on that hacking this problem requires solving the original problem of swapped instructions. I could have just done that. It took 9 milliseconds.

Choose not to point out their RGB pit would be rather blueish and not much else, given logical inference that every RGB code is meant to be #0000[0-F][0-3]

Also choose not to point out that this lagoon is about a thousand times larger than the surface area of the entire Earth. While we are not *on* Earth, I am guessing this lagoon by itself could envelop it from far enough away that the moon would fit too.

Leap from here to the previous island to save time.

It no longer hurts.

Close bag.