Day "19": "Solve" day 18
Attempt to educate elves on that, in the process of solving the problem this way, I have to reduce it to a trivial problem per part. If checking a part with four ints against four ranges of ints takes more than a nanosecond, maybe get a decent system. If we suppose a part must be moved 1 meter depending on its accepted/rejected status, then if you are moving it at less than 1/3 the speed of light, your shitty elven paws are the true holdup, not the part check. Which you can't do to begin with because guess what, the last thing before the meaningless count of acceptable part coordinates is my AcceptedRanges so you'd have to come crawling back anyway.

Hope that none of my current co-workers read these commit messages.

It no longer hurts.

Close bag.